Factor out the default package set into a packagegroup
Create a packagegroup (and required infrastructure) for a
recommended default package set.

The idea behind is that a packagegroup can be easier
modified by derived layers.

Note a small difference: the packagegroup now pulls nymea-data
which wasn't the case for the image. And libavahi-client is dropped
since nymea-zeroconf-plugin-avahi pulls it.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>